% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-svgz.R
\name{compare_file_text_svg}
\alias{compare_file_text_svg}
\title{Compare an SVG file that may be compressed}
\usage{
compare_file_text_svg(old, new)
}
\arguments{
\item{old}{character; path to SVG file to be used as reference}

\item{new}{character; path to SVG file to compare with \code{old}}
}
\value{
A logical vector of length 1.
}
\description{
Compare an SVG file that may be compressed
}
