% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-svgz.R
\name{read_svg}
\alias{read_svg}
\title{Read an SVG file that may be compressed}
\usage{
read_svg(f)
}
\arguments{
\item{f}{character; path to SVG file to be read.}
}
\value{
A character vector of length the number of lines read.
}
\description{
Read an SVG file that may be compressed
}
